# docs/swagger.yaml
basePath: /
definitions:
  handler.shortenUrlRequest:
    properties:
      exp:
        type: integer
      url:
        type: string
    required:
    - exp
    - url
    type: object
  model.User:
    properties:
      _:
        type: string
      display_name:
        type: string
      email:
        type: string
      id:
        type: string
      username:
        type: string
    type: object
  user.registerInput:
    properties:
      display_name:
        type: string
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - display_name
    - email
    - password
    - username
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for bookmark management
  title: Bookmark API
  version: "1.5"
paths:
  /health-check:
    get:
      description: ping and pong with redis server
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: check redis health
      tags:
      - health check
  /v1/links/redirect/{code}:
    get:
      consumes:
      - application/json
      parameters:
      - description: code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Found
      summary: Redirect Forward the request to the original url
      tags:
      - link
  /v1/links/shorten:
    post:
      consumes:
      - application/json
      parameters:
      - description: Shorten URL Input payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.shortenUrlRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: receive the url, return the code
      tags:
      - link
  /v1/users/register:
    post:
      consumes:
      - application/json
      description: Create a new user
      parameters:
      - description: User registration input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/user.registerInput'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            properties:
              data:
                $ref: '#/definitions/model.User'
              message:
                type: string
            type: object
      summary: Create a new user
      tags:
      - user
swagger: "2.0"